+++ Makefile
@@ -1,0 +1,11 @@
+.PHONY: install clean
+
+install:
+	uv sync --force-reinstall
+
+clean:
+	rm -rf build dist src/*.egg-info
+	rm -f src/waterz/*.so src/waterz/*.pyd src/waterz/evaluate.cpp
+
+test:
+	uv run pytest
