+++ README.md
@@ -1,42 +1,49 @@
+# waterz
+
 [![License](https://img.shields.io/pypi/l/waterz.svg?color=green)](https://github.com/funkey/waterz/raw/master/LICENSE)
 [![PyPI](https://img.shields.io/pypi/v/waterz.svg?color=green)](https://pypi.org/project/waterz)
 [![Python Version](https://img.shields.io/pypi/pyversions/waterz.svg?color=green)](https://python.org)
 [![CI](https://github.com/funkey/waterz/actions/workflows/test.yml/badge.svg)](https://github.com/funkey/waterz/actions/workflows/test.yml)
 [![codecov](https://codecov.io/gh/funkey/waterz/branch/main/graph/badge.svg?token=qGnz9GXpEb)](https://codecov.io/gh/funkey/waterz)
 
-<!--
-    the coerage is not working on c++ code, so the coverage is 0 now!
-    [![Coverage Status](https://coveralls.io/repos/github/jingpengw/waterz/badge.svg?branch=master)](https://coveralls.io/github/jingpengw/waterz?branch=master)
--->
-
-# waterz
-
 Pronounced *water-zed*. A simple watershed and region agglomeration library for
 affinity graphs.
 
-Based on the watershed implementation of [Aleksandar Zlateski](https://bitbucket.org/poozh/watershed) and [Chandan Singh](https://github.com/TuragaLab/zwatershed).
+Based on the watershed implementation of [Aleksandar
+Zlateski](https://bitbucket.org/poozh/watershed) and [Chandan
+Singh](https://github.com/TuragaLab/zwatershed).
 
-# Installation
-
-Install c++ dependencies:
-
-```sh
-sudo apt install libboost-dev
-```
-
-Install from PyPI
+## Install from PyPI
 
 ```sh
 pip install waterz
 ```
 
-install from local version
+## Install locally
+
+Install c++ dependencies:
 
 ```sh
-pip install .
+# linux
+sudo apt install libboost-dev
+# macos
+brew install boost
 ```
 
-# Usage
+Then
+
+```sh
+# make and activate env then:
+pip install -e .
+```
+
+or
+
+```sh
+uv sync
+```
+
+## Usage
 
 ```python
 import waterz
@@ -50,9 +57,9 @@ thresholds = [0, 100, 200]
 segmentations = waterz.agglomerate(affinities, thresholds)
 ```
 
-# Development
+## Development
 
-## Release to pypi
+### Release to pypi
 
 upgrade the version number in the `pyproject.toml` file, then
